Move johndoe over
