+++ cypress/user-data.ts
@@ -1,15 +1,3 @@
-export const johnDoe = {
-  email: "john.doe@kalle.app",
-  fullName: "John Doe",
-  password: "supersecurepassword",
+import { johnDoe as jd } from "../db/seeds"
 
-  calendars: {
-    baikal: {
-      name: "Baikal",
-      caldavAddress: "http://localhost:5232/dav.php/calendars/john.doe/test/",
-      username: "john.doe",
-      password: "root",
-      encryptedPassword: "ef87:y/u5eQgq3kbwuH8zeP/DVQ==:k3egrw==",
-    },
-  },
-}
+export const johnDoe = jd
