+++ db/seeds.ts
@@ -1,5 +1,20 @@
-import { johnDoe } from "cypress/user-data"
 import db from "db"
+
+export const johnDoe = {
+  email: "john.doe@kalle.app",
+  fullName: "John Doe",
+  password: "supersecurepassword",
+
+  calendars: {
+    baikal: {
+      name: "Baikal",
+      caldavAddress: "http://localhost:5232/dav.php/calendars/john.doe/test/",
+      username: "john.doe",
+      password: "root",
+      encryptedPassword: "ef87:y/u5eQgq3kbwuH8zeP/DVQ==:k3egrw==",
+    },
+  },
+}
 
 function mockHashPassword(pw: string) {
   if (pw === johnDoe.password) {
